Add a story for the badge
Need to remove the index /Badge logic from the Badge folder for it to work
