+++ src/stories/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { storiesOf, action, linkTo } from '@kadira/storybook';
-import Button from './Button';
+// import Alert from '../components/Alert';
+import Badge from '../components/Badge';
 import Welcome from './Welcome';
 
 storiesOf('Welcome', module)
@@ -8,10 +9,18 @@ storiesOf('Welcome', module)
     <Welcome showApp={linkTo('Button')}/>
   ));
 
-storiesOf('Button', module)
-  .add('with text', () => (
-    <Button onClick={action('clicked')}>Hello Button</Button>
-  ))
-  .add('with some emoji', () => (
-    <Button onClick={action('clicked')}>😀 😎 👍 💯</Button>
+storiesOf('Badge', module)
+  .add('with some text', () => (
+    <Badge>Hola</Badge>
   ));
+
+// storiesOf('Alert', module)
+//   .add('error', () => (
+//     <Alert isOpen type="error" heading="example alert" message="some error" />
+//   ))
+//   .add('warning', () => (
+//     <Alert isOpen type="warning" heading="example alert" message="something important" />
+//   ))
+//   .add('with onRemove', () => (
+//     <Alert isOpen type="error" heading="example alert" message="something important" onRemove={action('removed')} />
+//   ));
